lnwallet: move tests to input package
